refactor(slice-machine): convert SideNav to CSS Modules
