+++ packages/slice-machine/src/components/SideNav/SideNavEnvironmentSelector.module.css
@@ -1,0 +1,55 @@
+.logo {
+  display: block;
+  height: 40px;
+  width: 40px;
+}
+
+.environmentDotBase {
+  border-radius: 50%;
+  flex-shrink: 0;
+  height: 10px;
+  width: 10px;
+}
+
+.environmentDot-prod {
+  composes: environmentDotBase;
+  background-color: var(--purple10);
+}
+
+.environmentDot-stage {
+  composes: environmentDotBase;
+  background-color: var(--indigo10);
+}
+
+.environmentDot-dev {
+  composes: environmentDotBase;
+  background-color: var(--amber10);
+}
+
+.menuItemEnvironmentDot {
+  margin-left: 2px;
+  margin-right: 2px;
+}
+
+.activeEnvironmentDot {
+  bottom: 0;
+  outline-color: var(--grey2);
+  outline-style: solid;
+  outline-width: 3px;
+  position: absolute;
+  right: 0;
+  transform: translate(50%, 50%);
+}
+
+.activeEnvironmentName {
+  font-weight: 500;
+  line-height: 1.25rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.loginIcon {
+  height: 24px;
+  width: 24px;
+}
